requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dlib==19.24.9
face-recognition==1.3.0
face_recognition_models==0.3.0
idna==3.10
numpy==2.2.6
opencv-contrib-python==4.11.0.86
pandas==2.2.3
pillow==11.2.1
pyaes==1.6.1
pygame==2.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
